+++ controlplane/controller/internal/controller/metrics.go
@@ -33,7 +33,7 @@ var (
 		Name: "controller_grpc_getconfig_requests_total",
 		Help: "The total number of getconfig requests",
 	},
-		[]string{"pubkey", "device_code", "contributor_code", "exchange_code", "location_code", "agent_version", "agent_commit", "agent_date"},
+		[]string{"pubkey", "device_code", "contributor_code", "exchange_code", "location_code", "device_status", "agent_version", "agent_commit", "agent_date"},
 	)
 
 	getConfigMsgSize = prometheus.NewHistogram(prometheus.HistogramOpts{